+++ src/DataContainer/EntityImportSourceContainer.php
@@ -15,6 +15,7 @@ use Contao\System;
 use HeimrichHannot\EntityImportBundle\Event\AddSourceFieldMappingPresetsEvent;
 use HeimrichHannot\EntityImportBundle\Source\AbstractFileSource;
 use HeimrichHannot\EntityImportBundle\Source\CSVFileSource;
+use HeimrichHannot\EntityImportBundle\Source\RSSFileSource;
 use HeimrichHannot\EntityImportBundle\Source\SourceFactory;
 use HeimrichHannot\EntityImportBundle\Util\EntityImportUtil;
 use HeimrichHannot\UtilsBundle\Database\DatabaseUtil;
@@ -45,10 +46,12 @@ class EntityImportSourceContainer
 
     const FILETYPE_CSV = 'csv';
     const FILETYPE_JSON = 'json';
+    const FILETYPE_RSS = 'rss';
 
     const FILETYPES = [
         self::FILETYPE_CSV,
         self::FILETYPE_JSON,
+        self::FILETYPE_RSS,
     ];
 
     protected $activeBundles;
@@ -142,7 +145,6 @@ class EntityImportSourceContainer
 
                 switch ($fileType) {
                     case static::FILETYPE_CSV:
-
                         /** @var CSVFileSource $source */
                         $source = $this->sourceFactory->createInstance($sourceModel->id);
 
@@ -167,6 +169,20 @@ class EntityImportSourceContainer
 
                     case static::FILETYPE_JSON:
                         $dca['fields']['fileContent']['eval']['rte'] = 'ace|json';
+
+                        break;
+
+                    case static::FILETYPE_RSS:
+                        /** @var RSSFileSource $source */
+                        $source = $this->sourceFactory->createInstance($sourceModel->id);
+
+                        $options = $source->getPostFieldsAsOptions();
+
+                        $this->util->transformFieldMappingSourceValueToSelect(
+                            array_combine($options, $options)
+                        );
+
+                        $dca['fields']['fileContent']['eval']['rte'] = 'ace|xml';
 
                         break;
 
@@ -211,14 +227,19 @@ class EntityImportSourceContainer
         /** @var AbstractFileSource $source */
         $source = $this->sourceFactory->createInstance($dc->id);
 
-        if ($sourceModel->fileType === static::FILETYPE_CSV) {
-            return $source->getLinesFromFile(25, true)."\n...";
-        }
+        switch ($sourceModel->fileType) {
+            case static::FILETYPE_CSV:
+                return $source->getLinesFromFile(25, true)."\n...";
 
-        if ($sourceModel->fileType === static::FILETYPE_JSON) {
-            $string = json_decode($source->getFileContent(false));
+            case static::FILETYPE_JSON:
+                $string = json_decode($source->getFileContent(true));
 
-            return json_encode($string, JSON_PRETTY_PRINT);
+                return substr(json_encode($string, JSON_PRETTY_PRINT), 0, 50000);
+
+            case static::FILETYPE_RSS:
+                return substr($source->getFileContent(true), 0, 50000);
+
+                break;
         }
 
         return $source->getFileContent(true);