+++ src/Resources/contao/dca/tl_entity_import_source.php
@@ -86,6 +86,7 @@ $GLOBALS['TL_DCA']['tl_entity_import_source'] = [
         'retrievalType_absolute_path'      => 'absolutePath',
         'fileType_csv'                     => 'fileContent,csvHeaderRow,csvDelimiter,csvEnclosure,csvEscape,fieldMappingCopier,fieldMappingPresets,fieldMapping',
         'fileType_json'                    => 'fileContent,pathToDataArray,fieldMappingCopier,fieldMappingPresets,fieldMapping',
+        'fileType_rss'                     => 'fileContent,pathToDataArray,fieldMappingCopier,fieldMappingPresets,fieldMapping',
     ],
     // Fields
     'fields'      => [